Rewrite functions to take options as JSON object
Instead of functions having a long list of arguments, several of which
could be null, only require those arguments which are directly required
by Slack incoming webhooks themselves. The final argument is passed as a
JSON object containing any other options you want to use for the message
that you're posting to the webhook.